isolate spinner
