+++ src/Preloader.js
@@ -1,34 +1,6 @@
 import React from 'react';
 import cx from 'classnames';
-
-const Spinner = (props) => {
-  let {color, only} = props;
-  let spinnerClasses = {
-    'spinner-layer': true
-  };
-  if (only) {
-    spinnerClasses['spinner-' + color + '-only'] = true;
-  } else {
-    spinnerClasses['spinner-' + color] = true;
-  }
-  return (
-    <div className={cx(spinnerClasses)}>
-      <div className="circle-clipper left">
-        <div className="circle"></div>
-      </div>
-      <div className="gap-patch">
-        <div className="circle"></div>
-      </div>
-      <div className="circle-clipper right">
-        <div className="circle"></div>
-      </div>
-    </div>
-  );
-}
-
-Spinner.defaultProps = {
-  only: true,
-};
+import Spinner from './Spinner';
 
 let colors = ['blue', 'red', 'yellow', 'green'];
